Refactor: GetstartedCard reuseable components seperated from index
